move raw data helpers to utils
